Stable QStackedLayout across multiple files
3 different pages (3 files) are access from a QWidget class (one file) that is the central widget (`.setCentralWidget()`) of a GUI background class (one file).
